Make decompose module private in 0.2 version
