--- doc/source/apidoc/qutip_qip.decompose.rst
@@ -1,15 +1,0 @@
-﻿:orphan:
-
-qutip\_qip.decompose
-====================
-
-.. automodule:: qutip_qip.decompose
-   :members:
-   :show-inheritance:
-   :imported-members:
-
-   .. rubric:: Functions
-
-   .. autosummary::
-
-      decompose_one_qubit_gate
