+++ src/qutip_qip/_decompose/__init__.py
