+++ src/qutip_qip/_decompose/_utility.py
